Adding task ownership using authorization

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import {Task} from "./Task.entity";
 import {CreateTaskDto} from "./dto/create-task.dto";
 import {TaskStatus} from "./task.status.enum";
 import {GetTasksFilterDto} from "./dto/get-tasks-filter.dto";
+import {User} from "../auth/User.entity";
 
 @Injectable()
 export class TasksService {
@@ -14,13 +15,13 @@ export class TasksService {
         private taskRepository: TaskRepository) {
     }
 
-    async getTask(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTask(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
 
-    async getTaskById(id: number): Promise<Task> {
-        const foundTask = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const foundTask = await this.taskRepository.findOne({where: {id, userId: user.id}});
         if (!foundTask) {
             throw new NotFoundException(`Task with id: ${id} not found`);
         }
@@ -28,20 +29,20 @@ export class TasksService {
         return foundTask;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
         if (result.affected === 0) {
             throw new NotFoundException(`Task with id: ${id} not found`);
         }
 
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
